move dynamic registration
